+++ lib/openssl/keylog.rb
@@ -1,9 +1,0 @@
-require 'openssl'
-
-module OpenSSL
-  module Keylog
-    require 'openssl/keylog/ssl_socket_extensions'
-
-    OpenSSL::SSL::SSLSocket.send(:include, OpenSSL::Keylog::SSLSocketExtensions)
-  end
-end