+++ lib/sslkeylog.rb
@@ -1,0 +1,6 @@
+require 'openssl'
+
+module SSLkeylog
+  require 'sslkeylog/version'
+  require 'sslkeylog/openssl'
+end