+++ spec/spec_helper.rb
@@ -1,3 +1,3 @@
 $LOAD_PATH.unshift File.expand_path('../../lib', __FILE__)
-require 'openssl/keylog'
+require 'sslkeylog'
 require_relative 'shared/ssl_server_context.rb'